Expose the python package names, and link to kio, karchive etc
This string with python package names is used in places where the
package names are not the way they are on debian. E.g. opensuse has
slightly different naming convensions, and so we need to deal with
that. Without hardcoding things or keeping internal lists, this is
basically the only sane way i could come up with. Works just fine,
though, so it's not really a problem :)