Add install methods for Python and ZeroNet, and some feedback
This uses KIO and KArchive to download and decompress ZeroNet to
the location the user has set in configuration dialog.
It also uses PackageKit to install the required Python packages.
Further, it checks the system to see whether the two are already
installed.
Finally, it will provide feedback on what is currently going on,
by way of a workingOn property, which is empty when things are
not being worked on, and has a descriptive string when they are.